% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunay_graph.R
\name{delaunay_graph}
\alias{delaunay_graph}
\title{Build and filter a delaunay graph}
\usage{
delaunay_graph(las, global_filter = 0.8, max_edge_length = NULL)
}
\arguments{
\item{las}{aLAS file.}

\item{global_filter}{numeric between 0 and 1 (optional). If declared, applies a global
edge filter by removing all edges longer than the \eqn{global_filter^{th}}
quantile of all edges length.}

\item{max_edge_length}{numeric (optional). If declared, removes all edges longer than \emph{max_edge_length}.}
}
\value{
an data.table containing the nodes ID (their position within the LAS data) and the length of the edge.
}
\description{
Build and filter a delaunay graph
}
\examples{
\donttest{
# import data
file = system.file("extdata", "tree_no_leaves.las", package="lidUrb")
las = lidR::readLAS(file)

# build the knn graph with no filter
delaunay = lidUrb::delaunay_graph(las)

# plot the graph
lidUrb::plot_graph(las = las, graph = delaunay)
}
}
