% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_noise.R
\name{filter_noise}
\alias{filter_noise}
\title{Filter noise from TLS scene}
\usage{
filter_noise(las, k = 6L, sigma = 2, keep_noise = FALSE)
}
\arguments{
\item{las}{a LAS class object.}

\item{k}{numeric. the number of nearest neighbors to compute distance.}

\item{sigma}{numeric. The multiplier of the distance standard deviation to
class points as noise.}

\item{keep_noise}{logical. If TRUE the points classified as noise are kept in
the point cloud and are labeled in the is.noise fields,
if FALSE points classified as noise are removed.}
}
\value{
the LAS with a new field specifying if the points are noise (if keep_nois is TRUE) or
        the LAS without noise points (if keep_noise is FALSE).
}
\description{
Filter noise from TLS scene
}
\examples{
\donttest{
# import data
file = system.file("extdata", "urban.las", package="lidUrb")
las = lidR::readLAS(file)

# filter noise and store it for ploting
las = lidUrb::filter_noise(las,keep_noise = TRUE)

# plot noise in red
lidR::plot(las,color="is.noise")
}
}
