% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_features.R
\name{geom_features}
\alias{geom_features}
\title{Computes local geometrics features using a constant research distance}
\usage{
geom_features(las, search_radius, features_list)
}
\arguments{
\item{las}{a LAS file.}

\item{search_radius}{numeric. The search distance to retrieve point neighborhood.}

\item{features_list}{(optional) character. A vector containing the list of the geometric features to compute.
Can be: "Eigenvalue_sum","Omnivariance", "Eigenentropy","Anisotropy","Planarity",
"Linearity","PCA1","PCA2","Surface_variation","Sphericity","Verticality",
"Nx","Ny","Nz".}
}
\value{
The LAS file with the geometric features added as new columns in the slot data. NOTE: the names
        of the features is a contraction of the feature name and search_radius.
}
\description{
Compute local geometric features of the point cloud based on the eigenvalues of
             the points covariance matrix. This function uses the python implementation provided
             in \href{https://jakteristics.readthedocs.io/en/latest/}{jakteristics}.
}
\examples{
\donttest{
# import data
file = system.file("extdata", "urban.las", package="lidUrb")
las = lidR::readLAS(file)

# compute all features
las = lidUrb::geom_features(las,search_radius = 0.1)

# plot Planarity and Linearity
lidR::plot(las,color="Planarity")
lidR::plot(las,color="Linearity")
}
}
