% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clustering.R
\name{graph_clustering}
\alias{graph_clustering}
\title{Cluster data base on k nearest neighbor graph. Non connected components are
parts of different objects.}
\usage{
graph_clustering(las, graph)
}
\arguments{
\item{las}{a LAS file. NOTE: the data must be ordered similarly than when the graph was created.}

\item{graph}{a graph produced with the \code{\link{knn_graph}} or \code{\link{delaunay_graph}} functions.}
}
\value{
a vector with the points cluster
}
\description{
Cluster data base on k nearest neighbor graph. Non connected components are
parts of different objects.
}
\examples{
\donttest{
# import data
file = system.file("extdata", "urban.las", package="lidUrb")
las = lidR::readLAS(file)

# remove the ground
las@data = las@data[Classification == 0]

# build the knn graph with no filter
KNN = lidUrb::knn_graph(las, k = 5L)

# clustering
las = lidUrb::graph_clustering(las = las, graph = KNN)

# add random color to clusters
las@data[,cluster_plot := sample(1:20,1),by=cluster_graph]

# plot
lidR::plot(las,color="cluster_plot")
}
}
