% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsm_topology.R
\name{qsm_topology}
\alias{qsm_topology}
\title{Compute topology from a qsm file}
\usage{
qsm_topology(QSM_file)
}
\arguments{
\item{QSM_file}{a qsm file}
}
\value{
the qsm file now includin axid_ID, axis sections and branching order
}
\description{
Compute topology from a qsm file
}
\examples{
\donttest{
# import the obj file output from adTree
file = system.file("extdata", "adtree_qsm.obj", package="lidUrb")
mesh = Morpho::obj2mesh(file)

# transform the adTree mesh into a qsm file, simplify the mesh and compute
# some woody structure features
qsm = lidUrb::adtree2qsm(mesh)

# compute qsm topology
qsm$qsm = lidUrb::qsm_topology(qsm$qsm)

rgl::open3d()
rgl::shade3d(qsm$mesh,col=rep(qsm$qsm$branching_order,each=20),meshColor = "faces",add=T)

rgl::open3d()
rgl::shade3d(qsm$mesh,col=rep(qsm$qsm$axis_ID,each=20),meshColor = "faces",add=T)
}
}
