% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_point_density.R
\name{reduce_point_density}
\alias{reduce_point_density}
\title{Reduce the point density by keeping the point the closest to the center of a
voxel of given resolution}
\usage{
reduce_point_density(las, res, keep_points = FALSE)
}
\arguments{
\item{las}{a LAS class object.}

\item{res}{numeric. The voxel resolution.}

\item{keep_points}{logical. Allow to keep all the points in the point cloud
and label the reduced density points as \code{is.low.density == 1}
as well as the grpou of points each point belongs to (allow easy matching
after running a function on the downsampled point cloud).}
}
\value{
the LAS file with lower reduced density.
}
\description{
Reduce the point density by keeping the point the closest to the center of a
voxel of given resolution
}
\examples{
\donttest{
file = system.file("extdata", "urban.las", package="lidUrb")
las = lidR::readLAS(file)

# reduce point density with large voxels
las = lidUrb::reduce_point_density(las,0.2)

# plot
lidR::plot(las)
}
}
